%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Wed Feb 08 14:38:05 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{addCount.BigBang}
\alias{addCount.BigBang}
\alias{BigBang.addCount}
\alias{addCount.BigBang}
\alias{addCount,BigBang-method}
\alias{addCount}
\keyword{methods}
\keyword{internal}


 \title{Add a chromosome to rank and frequency stability counting}

 \description{
  Add a chromosome to the rank and frequency stability counting.
  This is an internal function
 }

 \usage{\method{addCount}{BigBang}(o, chr, ...)}

 \arguments{
	\item{chr}{Chromosome}
 }

 \value{
  Nothing.
 }

 \author{Victor Trevino. Francesco Falciani Group. University of Birmingham, U.K. http://www.bip.bham.ac.uk/bioinf}

 \seealso{
   For more information see \code{\link{BigBang}}.
 }

 \keyword{methods}
