%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Wed Feb 08 14:37:27 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{length.Bag}
\alias{length.Bag}
\alias{Bag.length}
\alias{length.Bag}
\alias{length,Bag-method}
\keyword{methods}
\keyword{internal}


\title{Gets the length of the object as its list version}

\description{
 Gets the length of the object as its list version.
}

\usage{\method{length}{Bag}(x, ...)}

\value{
 Returns length of the object.
}

\examples{
  b <- Bag(a=1,b=2,c=3)
  length(b)
}

\author{Victor Trevino. Francesco Falciani Group. University of Birmingham, U.K. http://www.bip.bham.ac.uk/bioinf}

\seealso{
  For more information see \code{\link{Bag}}.
}

\keyword{methods}
