%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Wed Feb 08 14:37:51 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{newRandomCollection.World}
\alias{newRandomCollection.World}
\alias{World.newRandomCollection}
\alias{newRandomCollection.World}
\alias{newRandomCollection,World-method}
\alias{newRandomCollection}
\keyword{methods}
\keyword{internal}


 \title{Creates a list of cloned object with its internal values generated by random}

 \description{
  Creates a list of cloned object with its internal values generated by random.
 }

 \usage{\method{newRandomCollection}{World}(.O, ...)}

 \arguments{
   \item{n}{Number of object clones.}
 }

 \value{
  Returns a list with cloned objects and random generated values.
 }

 \details{
  . For all cloned objects, \code{generateRandom} method is called.
 }

 \examples{
   cr <- Chromosome(genes=newCollection(Gene(shape1=1, shape2=100),5))
   cr
 
   ni <- Niche(chromosomes = newRandomCollection(cr, 2))
   ni

   wo <- World(niches = newRandomCollection(ni, 2))
   wo
 }

 \references{Goldberg, David E. 1989 \emph{Genetic Algorithms in Search, Optimization and Machine Learning}. Addison-Wesley Pub. Co. ISBN: 0201157675}

 \author{Victor Trevino. Francesco Falciani Group. University of Birmingham, U.K. http://www.bip.bham.ac.uk/bioinf}

 \seealso{
   For more information see \code{\link{World}}.
   \code{\link{unObject}},
   \code{\link{as.list.Object}},
   \code{\link{newCollection.World}},
   \code{\link{Chromosome}},
   \code{\link{Niche}},
   \code{\link{World}}.
 }

 \keyword{methods}
