%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Wed Feb 08 14:37:27 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.Bag}
\alias{print.Bag}
\alias{Bag.print}
\alias{print.Bag}
\alias{print,Bag-method}
\keyword{methods}
\keyword{internal}


\title{Prints the representation of the Bag object}

\description{
 Prints the representation of the Bag object.
}

\usage{\method{print}{Bag}(x, ...)}

\value{
 Returns nothing.
}

\examples{
  b <- Bag(a=1,b=2,c=3)
  b
  summary(b)
  as.list(b)
  unObject(b)
}

\author{Victor Trevino. Francesco Falciani Group. University of Birmingham, U.K. http://www.bip.bham.ac.uk/bioinf}

\seealso{
  For more information see \code{\link{Bag}}.
}

\keyword{print}
