\name{runifInt}
\alias{runifInt}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Generation of random uniform integer values}
\description{
 random number of uniform distribution
}
\usage{
runifInt(.O, n, mn, mx)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{.O}{ Gene objct }
  \item{n}{ Number of random values to generate }
  \item{mn}{ Minimum value }
  \item{mx}{ Maximum value }
}
\value{
 A vector with random values drawn from a uniform distribution.
}
\author{Victor Trevino. Francesco Falciani Group. University of Birmingham, U.K. }
\seealso{ \code{\link{runif}} }
 \keyword{methods}
